Deletes SearchRepository
