+++ core/data/src/main/kotlin/com/wachon/spotiwrap/data/di/DataModule.kt
@@ -4,12 +4,10 @@ import com.wachon.spotiwrap.data.repository.ArtistsRepository
 import com.wachon.spotiwrap.data.repository.DefaultArtistsRepository
 import com.wachon.spotiwrap.data.repository.DefaultGenresRepository
 import com.wachon.spotiwrap.data.repository.DefaultPlaylistRepository
-import com.wachon.spotiwrap.data.repository.DefaultSearchRepository
 import com.wachon.spotiwrap.data.repository.DefaultTracksRepository
 import com.wachon.spotiwrap.data.repository.DefaultUserRepository
 import com.wachon.spotiwrap.data.repository.GenresRepository
 import com.wachon.spotiwrap.data.repository.PlaylistRepository
-import com.wachon.spotiwrap.data.repository.SearchRepository
 import com.wachon.spotiwrap.data.repository.TracksRepository
 import com.wachon.spotiwrap.data.repository.UserRepository
 import org.koin.dsl.module
@@ -19,6 +17,5 @@ val DataModule = module {
     single<TracksRepository> { DefaultTracksRepository(get(), get()) }
     single<ArtistsRepository> { DefaultArtistsRepository(get(), get()) }
     single<GenresRepository> { DefaultGenresRepository(get()) }
-    single<SearchRepository> { DefaultSearchRepository(get()) }
     single<PlaylistRepository> { DefaultPlaylistRepository(get(), get()) }
 }
