+++ features/recommender/src/main/kotlin/com/wachon/spotiwrap/features/recommender/domain/SearchArtist.kt
@@ -2,7 +2,7 @@ package com.wachon.spotiwrap.features.recommender.domain
 
 import com.wachon.spotiwrap.core.common.dispatchers.DispatcherProvider
 import com.wachon.spotiwrap.core.common.model.ArtistModel
-import com.wachon.spotiwrap.data.repository.SearchRepository
+import com.wachon.spotiwrap.data.repository.ArtistsRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOn
 
@@ -11,10 +11,10 @@ interface SearchArtistUseCase {
 }
 
 class SearchArtist(
-    private val searchRepository: SearchRepository,
+    private val artistsRepository: ArtistsRepository,
     private val dispatchers: DispatcherProvider
 ) : SearchArtistUseCase {
     override suspend fun invoke(query: String): Flow<List<ArtistModel>> {
-        return searchRepository.searchArtist(query).flowOn(dispatchers.background)
+        return artistsRepository.searchArtist(query).flowOn(dispatchers.background)
     }
 }
