+++ features/recommender/src/main/kotlin/com/wachon/spotiwrap/features/recommender/domain/SearchTrack.kt
@@ -2,7 +2,7 @@ package com.wachon.spotiwrap.features.recommender.domain
 
 import com.wachon.spotiwrap.core.common.dispatchers.DispatcherProvider
 import com.wachon.spotiwrap.core.common.model.TrackModel
-import com.wachon.spotiwrap.data.repository.SearchRepository
+import com.wachon.spotiwrap.data.repository.PlaylistRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOn
 
@@ -11,10 +11,12 @@ interface SearchTrackUseCase {
 }
 
 class SearchTrack(
-    private val searchRepository: SearchRepository,
+    private val playlistRepository: PlaylistRepository,
     private val dispatchers: DispatcherProvider,
 ) : SearchTrackUseCase {
+
     override suspend fun invoke(query: String): Flow<List<TrackModel>> {
-        return searchRepository.searchTrack(query).flowOn(dispatchers.background)
+        return playlistRepository.searchTrack(query).flowOn(dispatchers.background)
     }
+
 }
